pole_performance/requirements.txt
Django==1.9.1
coverage==3.7.1
dj-database-url==0.3.0
django-allauth==0.24.1
django-environ==0.3.0
django-extensions==1.6.1
django-paypal==0.3.1
psycopg2==2.5.4
model-mommy==1.2.6
mock==1.0.1
gunicorn==19.3.0
pycrypto==2.6.1
simple-crypt==4.1.7
django-braces==1.9.0
python-dateutil==2.5.3